requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asyncssh==2.19.0
atpublic==5.1
attrs==25.1.0
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==44.0.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.7
entrypoints==0.4
fastapi==0.115.8
filelock==3.17.0
flatten-dict==0.4.2
flufl.lock==8.1.0
frozenlist==1.5.0
fsspec==2025.2.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.0.0
grandalf==0.8
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
iterative-telemetry==0.0.10
kombu==5.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.4.2
numpy==2.2.0
nvidia-ml-py==12.570.86
omegaconf==2.3.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
scipy==1.14.1
scmrepo==3.3.10
semver==3.0.4
setuptools==75.8.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.1
starlette==0.45.3
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
zc.lockfile==3.0.post1
